refactored basis evaluation classes into seperate files
