+++ include/nektar_interface/expansion_looping/geom_to_expansion_builder.hpp
@@ -1,6 +1,6 @@
 #ifndef __FUNCTION_COUPLING_BASE_H_
 #define __FUNCTION_COUPLING_BASE_H_
-#include "particle_interface.hpp"
+#include "nektar_interface/particle_interface.hpp"
 #include <map>
 #include <memory>
 #include <neso_particles.hpp>
